feat(08-02): integrate cache into embedding pipeline
- Add EmbeddingCache field to IndexPipeline
- Initialize cache in IndexPipeline::new() with model name from config
- Model name: 'jinaai/jina-embeddings-v2-base-code' for JinaCode, repo name for FastEmbed, 'hash' for Hash backend
- Cache enabled/disabled via config.embedding_cache_enabled (default: true)
- Max cache size: 1GB with lazy LRU cleanup every 1000 misses

- Modify embed_and_build_vector_records() to check cache first:
  * Check cache for each symbol text
  * Only embed uncached texts in batch
  * Store new embeddings in cache
  * Merge cached and new embeddings

- Log cache statistics after indexing completes:
  * hits, misses, hit_rate percentage
  * Helps monitor cache effectiveness

Deviations: None - implementation follows plan exactly.
Note: Pre-existing server compilation error (RefCell<Send> issue) not caused by this change.

+++ src/storage/sqlite/queries/cache.rs
@@ -109,5 +109,5 @@ pub fn cleanup_cache(conn: &Connection, max_size_bytes: i64) -> Result<i64> {
         []
     )?;
 
-    Ok(deleted)
+    Ok(deleted as i64)
 }
